코드&리드미 전반적으로 수정

+++ 강의/07_0316목/test5.c
@@ -2,10 +2,9 @@
 
 int main()
 {
-	// ToDo: int랑 double이 왜 안될까?
-	float a, b;
+	double a, b;
 	printf("두 수를 입력하세요: ");
-	scanf("%f %f", &a, &b);
+	scanf("%lf %lf", &a, &b);
 
 	char c;
 	printf("계산 부호를 입력하세요(+, -, *, /): ");
@@ -16,18 +15,18 @@ int main()
 	switch (c)
 	{
 	case '+':
-		printf("%f\n", a + b);
+		printf("%lf\n", a + b);
 		break;
 	case '-':
-		printf("%f\n", a - b);
+		printf("%lf\n", a - b);
 		break;
 	case '*':
-		printf("%f\n", a * b);
+		printf("%lf\n", a * b);
 		break;
 	case '/':
 		if (b != 0)
 		{
-			printf("%f\n", a / b);
+			printf("%lf\n", a / b);
 		}
 		else
 		{
